move network interface stuff to separate folder
