+++ src/transport/batchedNetworkInterface.ts
